Move challenge creation method to user services
