+++ app/src/auth/hooks/useWalletConnect.js
@@ -2,7 +2,7 @@ import { useCallback } from 'react'
 import WalletConnect from '@walletconnect/client'
 import QRCodeModal from '@walletconnect/qrcode-modal'
 import { convertUtf8ToHex } from '@walletconnect/utils'
-import { createChallenge } from '$shared/modules/integrationKey/services'
+import { createChallenge } from '$shared/modules/user/services'
 import { ChallengeFailedError } from '$shared/errors/Web3'
 import { post } from '$shared/utils/api'
 import routes from '$routes'
